requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.3.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0 